Add .env file for backend

+++ Backend/src/common/cleanUp/cleanUpImages.local.js
@@ -5,11 +5,18 @@ import { UPLOADFILES } from "../constant/app.constant";
 export const cleanUpImages = async () => {
   try {
     const images = await fs.readdir(UPLOADFILES);
+    const now = Date.now();
 
     for (const image of images) {
       const filePath = path.join(UPLOADFILES, image);
-      await fs.unlink(filePath);
-      console.log(`Deleted: ${image}`);
+      const stats = await fs.stat(filePath);
+
+      const fileAgeInMinutes = (now - stats.mtimeMs) / (1000 * 60); // mtimeMs là "modified time" tính bằng milliseconds
+
+      if (fileAgeInMinutes > 10) {
+        await fs.unlink(filePath);
+        console.log(`Deleted: ${image}`);
+      }
     }
   } catch (err) {
     console.error("Cleanup error:", err.message);
